+++ hephaistos-data/lua/Filters/CombatPresentation.lua
@@ -12,3 +12,18 @@ end
 Hephaistos.SetAnimation[StartLavaPresentation] = function(params)
 	return Hephaistos.MatchAll(params, { Name = "LavaVignetteSpawner", DestinationId = ScreenAnchors.LavaVignette })
 end
+
+-- fix full screen alert overlay displace fx
+-- for all calls e.g. call, [Redacted] call, Theseus call and Cerberus interaction during [Redacted] fight
+local function replaceObstacleAlertFx(params)
+  return Hephaistos.MatchAll(params, { Name = "FullscreenAlertDisplace", Group = "FX_Displacement", DestinationId = ScreenAnchors.FullscreenAlertFxAnchor })
+end
+Hephaistos.SpawnObstacle[DoAssistPresentation] = replaceObstacleAlertFx
+Hephaistos.SpawnObstacle[DoHadesAssistPresentation] = replaceObstacleAlertFx
+Hephaistos.SpawnObstacle[DoFullSuperPresentation] = replaceObstacleAlertFx
+Hephaistos.SpawnObstacle[DoTheseusSuperPresentation] = replaceObstacleAlertFx
+
+local function filterAnimationAlertFx(params)
+  return Hephaistos.MatchAll(params, { Name = "LegendaryAspectSnow", DestinationId = ScreenAnchors.FullscreenAlertFxAnchor })
+end
+Hephaistos.CreateAnimation[DoCerberusAssistPresentation] = filterAnimationAlertFx