+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 -   Re-fix world map shown between regions (oops :D).
+-   Fix full screen displacement FX for various occasions (calls, Chaos interact, Hades speaking, etc.)
 
 ## [v0.9.0] - 2021-08-21